Refactor include path caching
